add zoom

+++ src/visualization/Screen.hpp
@@ -9,7 +9,7 @@ class Screen {
 public:
     Screen(int width, int height);
     void set_protein(Protein* protein);
-    void set_zoom_level(int zoom);
+    void set_zoom_level(float zoom);
     bool handle_input();
     void drawScreen();
 
@@ -17,7 +17,7 @@ private:
     int screen_width;
     int screen_height;
     float aspect_ratio;
-    int zoom_level;
+    float zoom_level;
     char* mScreen;
     Protein* data;
